BUG FIX: Errors occurring multicore futures could prevent further multicore futures from being created [#26]
This bug was caught by chance thanks to the CRAN Policy; I added limitation to use options(mc.cores=2) for tests/futureAssign_OP.R.  This caused the five multicore futures to be limited to two cores and therefore each new future had to wait for the previous ones to be done.  In part of doing this, value() on erronous futures were called.